Ring light

+++ src/modes/BaseMode.cpp
@@ -1,5 +1,5 @@
 #include "BaseMode.h"
 
-CBaseMode::CBaseMode(const uint8_t numLeds, const String name)
+CBaseMode::CBaseMode(const uint16_t numLeds, const String name)
 : numLeds(numLeds), name(name) {
 }
