+++ src/modes/BaseMode.h
@@ -9,11 +9,11 @@ class CBaseMode {
 
 protected:
     unsigned long tMillis;
-    const uint8_t numLeds;
+    const uint16_t numLeds;
     const String name;
 
 public:
-	CBaseMode(const uint8_t numLeds, const String name);
+	CBaseMode(const uint16_t numLeds, const String name);
     virtual void draw(CRGB *leds) {};
 
     const String getName() { return name; }
